refactor(play): consolidate multiplayer and singleplayer scenario handling into a single component
