+++ frontend/src/features/playScenario/PlayScenarioPage.jsx
@@ -1,7 +1,7 @@
 import { useContext, useEffect, useState } from "react";
 import { useParams, useHistory } from "react-router-dom";
 import axios from "axios";
-import toast from "react-hot-toast";
+import { toast } from "react-hot-toast";
 
 import AuthenticationContext from "context/AuthenticationContext";
 import { usePost } from "hooks/crudHooks";
@@ -9,11 +9,15 @@ import { usePost } from "hooks/crudHooks";
 import LoadingPage from "../status/LoadingPage";
 import PlayScenarioCanvas from "./PlayScenarioCanvas";
 import { applyStateOperations } from "../../components/StateVariables/stateOperations";
+import { filterResourcesByConditions } from "../../utils/stateConditionalEvaluator";
+import NotesDisplayCard from "./modals/NotesModal/NotesModal";
+import ResourcesModal from "./modals/ResourcesModal/ResourcesModal";
+import PlayPageSideButton from "./components/PlayPageSideButton/PlayPageSideButton";
 import ResourcesPanel from "./components/ResourcesPanel";
 
 const sceneCache = new Map();
 
-const navigate = async (
+const navigateSingleplayer = async (
   user,
   scenarioId,
   currentScene,
@@ -42,15 +46,59 @@ const navigate = async (
   };
 };
 
+const navigateMultiplayer = async (
+  user,
+  groupId,
+  currentScene,
+  addFlags,
+  removeFlags,
+  componentId
+) => {
+  const token = await user.getIdToken();
+  const config = {
+    method: "post",
+    url: `/api/navigate/group/${groupId}`,
+    headers: {
+      "Content-Type": "application/json",
+      Authorization: `Bearer ${token}`,
+    },
+    data: { currentScene, addFlags, removeFlags, componentId },
+  };
+  const res = await axios.request(config);
+  if (res.data.scenes) {
+    res.data.scenes.forEach((scene) => sceneCache.set(scene._id, scene));
+  }
+  return {
+    newSceneId: res.data.active,
+    stateVariables: res.data.stateVariables,
+    newStateVersion: res.data.stateVersion,
+  };
+};
+
+const getMultiplayerResources = async (user, groupId) => {
+  const token = await user.getIdToken();
+  const config = {
+    method: "get",
+    url: `/api/navigate/group/resources/${groupId}`,
+    headers: {
+      "Content-Type": "application/json",
+      Authorization: `Bearer ${token}`,
+    },
+  };
+  const res = await axios.request(config);
+  return res.data;
+};
+
 /**
  * This page allows users to play a scenario.
  *
  * @container
  */
-export default function PlayScenarioPage() {
+export default function PlayScenarioPage({ group }) {
   const { user, loading, error: authError } = useContext(AuthenticationContext);
   const { scenarioId } = useParams();
   const history = useHistory();
+  const isMultiplayer = history.location.pathname.includes("/multiplayer/");
 
   const [sceneId, setSceneId] = useState(null);
   const [stateVariables, setStateVariables] = useState([]);
@@ -59,6 +107,8 @@ export default function PlayScenarioPage() {
   const [removeFlags, setRemoveFlags] = useState([]);
 
   const [resourcesOpen, setResourcesOpen] = useState(false);
+  const [noteOpen, setNoteOpen] = useState(false);
+  const [resources, setResources] = useState([]);
 
   const currScene = sceneCache.get(sceneId);
 
@@ -67,8 +117,13 @@ export default function PlayScenarioPage() {
     if (error.status === 409) {
       onSceneChange();
       toast.success(
-        "A move from somewhere else was made, but you're back on track!"
+        isMultiplayer
+          ? "Someone else made a move first, but you're back on track!"
+          : "A move from somewhere else was made, but you're back on track!"
       );
+    } else if (isMultiplayer && error.status === 403) {
+      const roles = JSON.stringify(error.meta.roles_with_access);
+      history.push(`/play/${scenarioId}/invalid-role?roles=${roles}`);
     } else {
       history.push(`/play/${scenarioId}/error`);
     }
@@ -89,14 +144,32 @@ export default function PlayScenarioPage() {
     }
 
     try {
-      const { newSceneId, stateVariables, newStateVersion } = await navigate(
-        user,
-        scenarioId,
-        sceneId,
-        addFlags,
-        removeFlags,
-        componentId
-      );
+      const { newSceneId, stateVariables, newStateVersion } = isMultiplayer
+        ? await navigateMultiplayer(
+            user,
+            group._id,
+            sceneId,
+            addFlags,
+            removeFlags,
+            componentId
+          )
+        : await navigateSingleplayer(
+            user,
+            scenarioId,
+            sceneId,
+            addFlags,
+            removeFlags,
+            componentId
+          );
+
+      if (isMultiplayer) {
+        const newResources = await getMultiplayerResources(user, group._id);
+        const filteredResources = filterResourcesByConditions(
+          newResources,
+          stateVariables
+        );
+        setResources(filteredResources);
+      }
 
       if (stateVersion < newStateVersion) {
         setStateVariables(stateVariables);
@@ -130,49 +203,90 @@ export default function PlayScenarioPage() {
   };
 
   const reset = async () => {
+    const resetUrl = isMultiplayer
+      ? `api/navigate/group/reset/${group._id}`
+      : `api/navigate/user/reset/${scenarioId}`;
     const res = await usePost(
-      `api/navigate/user/reset/${scenarioId}`,
+      resetUrl,
       { currentScene: sceneId },
       user.getIdToken.bind(user)
     );
-    if (res.status) {
+
+    if (res?.status) {
       handleError(res);
       return;
     }
 
+    setAddFlags([]);
+    setRemoveFlags([]);
     onSceneChange();
   };
 
   if (loading) return <LoadingPage text="Loading Scene..." />;
   if (authError) return <></>;
+  if (isMultiplayer && !group) return <LoadingPage text="Loading Scene..." />;
   if (!currScene) return <LoadingPage text="Loading Scene..." />;
+
+  const incrementor = (id) => {
+    if (!sceneCache.has(id)) return;
+    history.replace(`/play/${scenarioId}/multiplayer/${id}`);
+  };
 
   return (
     <div className="w-full h-full relative">
       <PlayScenarioCanvas
         scene={currScene}
+        incrementor={isMultiplayer ? incrementor : undefined}
         reset={reset}
         setAddFlags={setAddFlags}
         setRemoveFlags={setRemoveFlags}
         buttonPressed={buttonPressed}
         stateVariables={stateVariables}
       />
-      <div className="absolute top-2 right-2 z-30 flex items-center gap-2">
-        <button
-          className="btn btn-sm"
-          onClick={() => setResourcesOpen(true)}
-          aria-label="Open resources"
-        >
-          Resources
-        </button>
-      </div>
 
-      <ResourcesPanel
-        scenarioId={scenarioId}
-        stateVariables={stateVariables}
-        open={resourcesOpen}
-        onClose={() => setResourcesOpen(false)}
-      />
+      {isMultiplayer ? (
+        <>
+          <PlayPageSideButton
+            setNoteOpen={setNoteOpen}
+            setResourcesOpen={setResourcesOpen}
+          />
+
+          {noteOpen && (
+            <NotesDisplayCard
+              group={group}
+              user={user}
+              handleClose={() => setNoteOpen(false)}
+            />
+          )}
+          {resourcesOpen && (
+            <ResourcesModal
+              resources={resources}
+              group={group}
+              user={user}
+              handleClose={() => setResourcesOpen(false)}
+            />
+          )}
+        </>
+      ) : (
+        <>
+          <div className="absolute top-2 right-2 z-30 flex items-center gap-2">
+            <button
+              className="btn btn-sm"
+              onClick={() => setResourcesOpen(true)}
+              aria-label="Open resources"
+            >
+              Resources
+            </button>
+          </div>
+
+          <ResourcesPanel
+            scenarioId={scenarioId}
+            stateVariables={stateVariables}
+            open={resourcesOpen}
+            onClose={() => setResourcesOpen(false)}
+          />
+        </>
+      )}
     </div>
   );
 }
